Completed Rock-Paper-Scissor project gameplay code
Finished the Rock-Paper-Scissors gameplay by using built in methods such as JSON and LocalStorage to cause the score to update after each time the player plays. And used LocalStorage to store the score permanently even when the user reloads the page. Converted JavaScript Object Notation to JSON and vice versa. Used Objects to clean the code by grouping and using code of similar topics.